% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLS.R
\name{GLS}
\alias{GLS}
\title{Title}
\usage{
GLS(y, X, sigma)
}
\arguments{
\item{y}{Response}

\item{X}{model matrix}

\item{sigma}{Uninverted covariance matrix}
}
\value{
GLS estimator of beta
}
\description{
Title
}
\examples{
y=c(0.978,1.08) x=(0.383,0.45) sigma=diag(2)
}
