% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{This is a description of linreg function}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A formula}

\item{data}{A matrix}
}
\value{
An object with of class linreg as an S3 class
}
\description{
This function creates a multiple regression model to handle linear regression.
}
\details{
Two arguments with a formula and a data are handled by this function.
}
\examples{
linreg(Petal.Length ~ Species, data = iris)

}
