\name{Regression-package}
\alias{Regression-package}
\alias{Regression}
\docType{package}
\title{
Fuctions for linear models
}
\description{
A (eventual) collection of tools for linear models
}
\details{
\tabular{ll}{
Package: \tab Regression\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-09-06\cr
License: \tab GPL-2\cr
}
}
\author{
Tyler Rinker

Maintainer: Tyler Rinker <tyler.rinker@gmail.com>
}