% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-model.R
\name{regression.fit}
\alias{regression.fit}
\title{Fits a least squares regression model}
\usage{
regression.fit(X, y)
}
\arguments{
\item{X}{Design matrix of input values}

\item{y}{Output values}
}
\value{
Least squares linear regression estimator
}
\description{
Fits a least squares regression model
}
\examples{

}
