% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-model.R
\name{regression.predict}
\alias{regression.predict}
\title{Predict output from input data}
\usage{
regression.predict(params, data)
}
\arguments{
\item{params}{Least squares estimator}

\item{data}{new input data}
}
\value{
Prediction of output values
}
\description{
Predict output from input data
}
