% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg}
\alias{linreg}
\title{A Reference Class to perform Linear Regression}
\description{
A Class to fit linear models. It can be used to carry out regression.
}
\section{Fields}{

\describe{
\item{\code{formula}}{An object of class formula}

\item{\code{data}}{A data frame from which variables are taken to formula}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{Returns the coefficient as a named vector}

\item{\code{get_beta_coef()}}{Returns the Beta Coefficient values}

\item{\code{get_df()}}{Returns the degree of freedom}

\item{\code{get_fitted_values()}}{Return the Predicted value for 'Dependent variable' }

\item{\code{get_ixtx()}}{Returns Inverse(T(X)*X) i.e., inverse of dot product of t(x) and x.}

\item{\code{get_res_var()}}{Returns Residual of Variance}

\item{\code{get_residuals()}}{Returns the residuals for the dependent variable(Y)}

\item{\code{get_t_beta()}}{Returns t-statistic for each beta coefficient}

\item{\code{get_var_beta()}}{Returns Variance for each Beta coefficient.}

\item{\code{initialize(formula, data, ..., use.qr = F)}}{Assign 'formula', 'data' and 'use.qr' upon object creation }

\item{\code{plot()}}{Plots the 'Residual vs Fitted' and 'Scale-Location' plots respectively }

\item{\code{pred()}}{Returns the predicted value y_hat}

\item{\code{print()}}{Prints out the Coefficients and Coefficient names}

\item{\code{resid()}}{Returns the vector of Residuals e}

\item{\code{summary()}}{Function to summarise Inferential statistics summary similar to summary.lm()}

\item{\code{yield_x()}}{Returns X(Independent Variables) along with the intercept}

\item{\code{yield_y()}}{Returns Y(dependent variable)}
}}

\examples{
lr_object <- linreg$new(formula = Sepal.Length~Petal.Width, data = iris)

lr_object$print()
lr_object$plot()
lr_object$resid()
lr_object$pred()
lr_object$coef()

}
\references{
\href{https://liu.se/insidan/kommunikationsstod/grafiskprofil?l=en}{LiU Theme}
}
